+++ src/components/layout/Sidebar/index.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+import MyPage from './MyPage';
+import NewPage from './NewPageButton';
+import Options from './Options';
+import SidebarLayout from './SidebarLayout';
+
+function Sidebar() {
+  return (
+    <SidebarLayout>
+      <Options />
+      <MyPage />
+      <NewPage />
+    </SidebarLayout>
+  );
+}
+
+export default Sidebar;
